+++ kai/ukernels/matmul/matmul_clamp_f32_f16p_qsi4c32p/kai_matmul_clamp_f32_f16p1vlx2_qsi4c32p4vlx2_1vlx4vl_sme2_mopa.h
@@ -60,7 +60,7 @@ size_t kai_get_sr_matmul_clamp_f32_f16p1vlx2_qsi4c32p4vlx2_1vlx4vl_sme2_mopa(voi
 ///
 /// @param[in] m_idx Row index in the LHS matrix (not packed). It must be a multiple of m_step.
 /// @param[in] k     Total number of columns in the LHS matrix (not packed).
-/// @param[in] bl    Block length. It must be 32.
+/// @param[in] bl    Block length. It must be a multiple of 32.
 ///
 /// @return the offset in bytes to the packed LHS matrix
 size_t kai_get_lhs_packed_offset_matmul_clamp_f32_f16p1vlx2_qsi4c32p4vlx2_1vlx4vl_sme2_mopa(
@@ -73,7 +73,7 @@ size_t kai_get_lhs_packed_offset_matmul_clamp_f32_f16p1vlx2_qsi4c32p4vlx2_1vlx4v
 ///
 /// @param[in] n_idx Row index in the RHS matrix (not packed). It must be a multiple of n_step.
 /// @param[in] k     The common dimension between the LHS and RHS matrix (K).
-/// @param[in] bl    Block length. It must be 32.
+/// @param[in] bl    Block length. It must be a multiple of 32.
 ///
 /// @return the offset in bytes to the packed RHS matrix
 size_t kai_get_rhs_packed_offset_matmul_clamp_f32_f16p1vlx2_qsi4c32p4vlx2_1vlx4vl_sme2_mopa(
@@ -114,7 +114,7 @@ size_t kai_get_dst_size_matmul_clamp_f32_f16p1vlx2_qsi4c32p4vlx2_1vlx4vl_sme2_mo
 /// @param[in]  m              The number of output rows written.
 /// @param[in]  n              The number of output columns written.
 /// @param[in]  k              The number of channels. The common dimension between the LHS and RHS matrix.
-/// @param[in]  bl             Block length. Block length. It must be 32.
+/// @param[in]  bl             Block length. It must be a multiple of 32.
 /// @param[in]  lhs_packed     The LHS packed matrix. The micro-kernel dependencies list at the top of this file reports
 ///                            the available LHS packing functions for this micro-kernel.
 /// @param[in]  rhs_packed     The RHS packed matrix. The micro-kernel dependencies list at the top of this file reports