+++ kai/ukernels/matmul/matmul_clamp_f32_qsi8d32p_qsi4c32p/kai_matmul_clamp_f32_qsi8d32p1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa.c
@@ -1,5 +1,5 @@
 //
-// SPDX-FileCopyrightText: Copyright 2024-2025 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: Copyright 2024-2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -43,16 +43,16 @@ inline static size_t kai_get_num_bytes_per_block_lhs(size_t bl) {
 }
 
 inline static size_t kai_get_num_bytes_per_block_rhs(size_t bl) {
-    KAI_ASSUME(bl == kai_bl);
+    KAI_ASSUME((bl % kai_bl) == 0);
     size_t num_bytes_per_block_rhs = (bl / kai_recip_num_bytes_qvalue_rhs) + kai_num_bytes_multiplier_rhs;
     return num_bytes_per_block_rhs;
 }
 
 inline static size_t kai_get_num_blocks_per_row(size_t k, size_t bl) {
-    KAI_ASSUME(bl == kai_bl);
-    KAI_ASSUME((k % kai_bl) == 0);
+    KAI_ASSUME((bl % kai_bl) == 0);
+    KAI_ASSUME((k % bl) == 0);
 
-    return kai_roundup(k, bl) / bl;
+    return k / bl;
 }
 
 inline static size_t kai_get_lhs_packed_stride(size_t k, size_t bl) {
@@ -61,8 +61,8 @@ inline static size_t kai_get_lhs_packed_stride(size_t k, size_t bl) {
 }
 
 inline static size_t kai_get_rhs_packed_stride(size_t k, size_t bl) {
-    KAI_ASSUME(bl == kai_bl);
-    KAI_ASSUME((k % kai_bl) == 0);
+    KAI_ASSUME((bl % kai_bl) == 0);
+    KAI_ASSUME((k % bl) == 0);
 
     const size_t num_blocks_per_row = kai_get_num_blocks_per_row(k, bl);
     const size_t num_bytes_per_block = kai_get_num_bytes_per_block_rhs(bl);
@@ -143,6 +143,7 @@ void kai_run_matmul_clamp_f32_qsi8d32p1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa(
     float scalar_min,                 //
     float scalar_max) {
     KAI_ASSUME(dst_stride_col == sizeof(float));
+    KAI_ASSUME((bl % kai_bl) == 0);
 
     KAI_UNUSED(scalar_min);
     KAI_UNUSED(scalar_max);
@@ -155,6 +156,7 @@ void kai_run_matmul_clamp_f32_qsi8d32p1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa(
         size_t lhs_packed_stride;
         size_t rhs_packed_stride;
         size_t mr;
+        size_t bl;
     } KernelArgs;
 
     KernelArgs ka;
@@ -167,6 +169,7 @@ void kai_run_matmul_clamp_f32_qsi8d32p1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa(
     ka.mr = mr;
     ka.lhs_packed_stride = kai_get_lhs_packed_stride(k, bl);
     ka.rhs_packed_stride = kai_get_rhs_packed_stride(k, bl);
+    ka.bl = bl;
 
     const uint16_t* lhs_scales = (const uint16_t*)((const int8_t*)lhs_packed + ka.lhs_packed_stride -
                                                    (mr * num_blocks) * kai_num_bytes_multiplier_lhs);
@@ -251,7 +254,8 @@ void kai_run_matmul_clamp_f32_qsi8d32p1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa(
         // Iterate over all values in the block
         // k_blk_idx = bl
         // e.g. while(k_blk_idx > 0) {... k_blk_idx -= 4}
-        " mov x11, #32\n"
+        " ldr x5, [%x[args_ptr], %[bl]] \n"
+        " mov x11, x5\n"
 
         " 4: // .LOOP_BL_START%=: \n"
 
@@ -346,7 +350,8 @@ void kai_run_matmul_clamp_f32_qsi8d32p1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa(
         " blt 5b // .LOOP_ZA%= \n"
 
         // Decrement K loop index by bl
-        " subs x10, x10, #32 \n"
+        " ldr x5, [%x[args_ptr], %[bl]] \n"
+        " subs x10, x10, x5 \n"
 
         " b.gt 3b // .LOOP_K_START%= \n"
 
@@ -398,7 +403,8 @@ void kai_run_matmul_clamp_f32_qsi8d32p1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa(
         : [M] "r"(m), [N] "r"(n), [K] "r"(k), [lhs_packed] "r"(lhs_packed), [lhs_scales] "r"(lhs_scales),
           [stride] "r"(dst_stride_row), [lut] "r"(lut), [args_ptr] "r"(&ka),
           [offset_stride_l] "I"(offsetof(KernelArgs, lhs_packed_stride)),
-          [offset_stride_r] "I"(offsetof(KernelArgs, rhs_packed_stride)), [offset_mr] "I"(offsetof(KernelArgs, mr))
+          [offset_stride_r] "I"(offsetof(KernelArgs, rhs_packed_stride)), [offset_mr] "I"(offsetof(KernelArgs, mr)),
+          [bl] "I"(offsetof(KernelArgs, bl))
         : "p0", "p1", "p3", "p4", "p5", "p6", "p7", "p8", "p9", "p10", "p11", "p12", "p13", "p14", "p15", "z0", "z1",
           "z2", "z3", "z4", "z5", "z6", "z7", "z8", "z9", "z10", "z11", "z12", "z13", "z14", "z15", "z16", "z17", "z18",
           "z19", "z20", "z21", "z22", "z23", "z24", "z25", "z26", "z27", "z28", "z29", "z30", "z31", "x0", "x5", "x6",