+++ test/tests/matmul_clamp_f32_qsi8d32p_qsi4c32p_test.cpp
@@ -1,5 +1,5 @@
 //
-// SPDX-FileCopyrightText: Copyright 2024-2025 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: Copyright 2024-2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -84,8 +84,8 @@ struct UKernelVariants {
 };
 
 // clang-format off
-static const int num_non_clamping_kernels = 4;
-static const std::array<UKernelVariants, 11>
+static const int num_non_clamping_kernels = 2;
+static const std::array<UKernelVariants, 9>
     variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p = {
         {
          // NOTE: The following kernels do not support clamping despite their names.
@@ -95,12 +95,6 @@ static const std::array<UKernelVariants, 11>
          {UKERNEL_MATMUL_PACK_VARIANT(
              clamp_f32_qsi8d32p1x4_qsi4c32p4x4_1x4_neon_dotprod, cpu_has_dotprod, lhs_quant_pack_qsi8d32p_f32,
              rhs_pack_nxk_qsi4c32pscalef16_qsu4c32s16s0, false), false},
-         {UKERNEL_MATMUL_PACK_VARIANT(
-             clamp_f32_qsi8d32p1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa, cpu_has_sme2, lhs_quant_pack_qsi8d32p_f32_neon,
-             rhs_pack_nxk_qsi4c32ps1s0scalef16_qsu4c32s16s0_neon, false), false},
-         {UKERNEL_MATMUL_PACK_VARIANT(
-             clamp_f32_qsi8d32p1x4_qsi4c32p4vlx4_1x4vl_sme2_sdot, cpu_has_sme2, lhs_quant_pack_qsi8d32p_f32_neon,
-             rhs_pack_nxk_qsi4c32ps1s0scalef16_qsu4c32s16s0_neon, false), false},
 
          // The kernels below this point will run clamping tests
          {UKERNEL_MATMUL_PACK_VARIANT(
@@ -124,10 +118,46 @@ static const std::array<UKernelVariants, 11>
          {UKERNEL_MATMUL_PACK_VARIANT(
               clamp_f32_qsi8d32p4x8_qsi4c32p8x8_16x8_sve_i8mm, (cpu_check<cpu_has_sve_vl256, cpu_has_i8mm>), lhs_quant_pack_qsi8d32p_f32,
               rhs_pack_nxk_qsi4c32pscalef16_qsu4c32s16s0, false), true}}};
+
+static const std::array<UKernelVariants, 2>
+    variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p_variable_bl = {
+        {
+         {UKERNEL_MATMUL_PACK_VARIANT(
+            clamp_f32_qsi8d32p1vlx4_qsi4c32p4vlx4_1vlx4vl_sme2_mopa, cpu_has_sme2, lhs_quant_pack_qsi8d32p_f32_neon,
+            rhs_pack_nxk_qsi4c32ps1s0scalef16_qsu4c32s16s0_neon, false), false},
+         {UKERNEL_MATMUL_PACK_VARIANT(
+            clamp_f32_qsi8d32p1x4_qsi4c32p4vlx4_1x4vl_sme2_sdot, cpu_has_sme2, lhs_quant_pack_qsi8d32p_f32_neon,
+            rhs_pack_nxk_qsi4c32ps1s0scalef16_qsu4c32s16s0_neon, false), false}}};
 // clang-format on
 
-class MatMulTest_f32_qsi8d32p_qsi4c32p
-    : public ::testing::TestWithParam<std::tuple<size_t, MatMulShape, MatrixPortion, float>> {};
+using MatMulTestParams_f32_qsi8d32p_qsi4c32p = std::tuple<size_t, MatMulShape, MatrixPortion, float, size_t, bool>;
+
+[[maybe_unused]] static void PrintTo(const MatMulTestParams_f32_qsi8d32p_qsi4c32p& param, std::ostream* os) {
+    const auto variant_idx = std::get<0>(param);
+    const auto shape = std::get<1>(param);
+    const auto portion = std::get<2>(param);
+    const auto clamp_keep_ratio = std::get<3>(param);
+    const auto bl = std::get<4>(param);
+    const auto variable_bl = std::get<5>(param);
+
+    *os << "variant_" << variant_idx << "__";
+    PrintTo(shape, os);
+    *os << "__";
+    PrintTo(portion, os);
+    *os << "__clamp_keep_ratio_" << static_cast<int>(clamp_keep_ratio * 100);
+    *os << (variable_bl ? "__VarBL" : "__FixedBL");
+    *os << "__bl_" << bl;
+}
+
+class MatMulTest_f32_qsi8d32p_qsi4c32p : public ::testing::TestWithParam<MatMulTestParams_f32_qsi8d32p_qsi4c32p> {};
+
+static const UKernelVariants& get_variant_entry(size_t variant_index, bool variable_bl) {
+    if (!variable_bl) {
+        return variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p.at(variant_index);
+    }
+
+    return variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p_variable_bl.at(variant_index);
+}
 
 // Ensure non-clamping tests are marked correctly.
 TEST(KernelClampingCheck, SanityCheck) {
@@ -137,17 +167,18 @@ TEST(KernelClampingCheck, SanityCheck) {
 }
 
 TEST_P(MatMulTest_f32_qsi8d32p_qsi4c32p, Offset_RHS) {
-    const auto& [variant_index, matmul_shape, portion, clamp_keep_ratio] = GetParam();
-    const auto& ukernel_variant = variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p.at(variant_index).variant;
+    const auto& [variant_index, matmul_shape, portion, clamp_keep_ratio, bl, variable_bl] = GetParam();
+    const auto& ukernel_variant = get_variant_entry(variant_index, variable_bl).variant;
 
     if (ukernel_variant.ukernel.fn_is_supported && !ukernel_variant.ukernel.fn_is_supported()) {
         GTEST_SKIP() << "Unsupported CPU feature";
     }
 
-    const size_t bl = 32;
     const size_t M = matmul_shape.m;
     const size_t N = matmul_shape.n;
     const size_t K = matmul_shape.k;
+
+    ASSERT_TRUE(K % bl == 0);
 
     const auto nr = ukernel_variant.ukernel.interface.get_nr();
     const auto kr = ukernel_variant.ukernel.interface.get_kr();
@@ -167,17 +198,18 @@ TEST_P(MatMulTest_f32_qsi8d32p_qsi4c32p, Offset_RHS) {
 }
 
 TEST_P(MatMulTest_f32_qsi8d32p_qsi4c32p, Offset_LHS) {
-    const auto& [variant_index, matmul_shape, portion, clamp_keep_ratio] = GetParam();
-    const auto& ukernel_variant = variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p.at(variant_index).variant;
+    const auto& [variant_index, matmul_shape, portion, clamp_keep_ratio, bl, variable_bl] = GetParam();
+    const auto& ukernel_variant = get_variant_entry(variant_index, variable_bl).variant;
 
     if (ukernel_variant.ukernel.fn_is_supported && !ukernel_variant.ukernel.fn_is_supported()) {
         GTEST_SKIP() << "Unsupported CPU feature";
     }
 
-    const size_t bl = 32;
     const size_t M = matmul_shape.m;
     const size_t N = matmul_shape.n;
     const size_t K = matmul_shape.k;
+
+    ASSERT_TRUE(K % bl == 0);
 
     const auto mr = ukernel_variant.ukernel.interface.get_mr();
     const auto kr = ukernel_variant.ukernel.interface.get_kr();
@@ -199,21 +231,22 @@ TEST_P(MatMulTest_f32_qsi8d32p_qsi4c32p, Offset_LHS) {
 }
 
 TEST_P(MatMulTest_f32_qsi8d32p_qsi4c32p, EndToEnd) {
-    const auto& [variant_index, matmul_shape, portion, clamp_keep_ratio] = GetParam();
-    const auto& ukernel_variant = variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p.at(variant_index).variant;
+    const auto& [variant_index, matmul_shape, portion, clamp_keep_ratio, bl, variable_bl] = GetParam();
+    const auto& ukernel_variant = get_variant_entry(variant_index, variable_bl).variant;
 
     if (ukernel_variant.ukernel.fn_is_supported && !ukernel_variant.ukernel.fn_is_supported()) {
         GTEST_SKIP() << "Unsupported CPU feature";
     }
 
     // NOTE: Workaround - some kernels despite being called matmul_clamp do not support clamping.
-    const bool clamp_support = variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p.at(variant_index).clamp_support;
+    const bool clamp_support = get_variant_entry(variant_index, variable_bl).clamp_support;
     const std::uint32_t seed = 0;
 
     const size_t M = matmul_shape.m;
     const size_t N = matmul_shape.n;
     const size_t K = matmul_shape.k;
-    const size_t bl = 32;
+
+    ASSERT_TRUE(K % bl == 0);
 
     const auto mr = ukernel_variant.ukernel.interface.get_mr();
     const auto nr = ukernel_variant.ukernel.interface.get_nr();
@@ -340,15 +373,17 @@ INSTANTIATE_TEST_SUITE_P(
             MatrixPortion(0, 0.75, 1, 1),  // Rightmost portion.
             MatrixPortion(0, 0.5, 1, 0.8)  // Somewhere Middle
             ),
-        testing::ValuesIn(std::initializer_list<float>{1.0F})),  // We keep 100% of values - no clamping
+        testing::ValuesIn(std::initializer_list<float>{1.0F}),  // We keep 100% of values - no clamping
+        testing::Values(32), testing::Values(false)),
     [](const auto& info) {
         const auto variant_idx = std::get<0>(info.param);
         const std::string name{variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p.at(variant_idx).variant.ukernel.name};
         const auto shape = std::get<MatMulShape>(info.param);
         const auto portion = std::get<2>(info.param);
         const auto clamp_keep_ratio = std::get<3>(info.param);
+        const auto bl = std::get<4>(info.param);
 
-        return test_description(name, shape, portion, true, clamp_keep_ratio);
+        return test_description(name, shape, portion, true, clamp_keep_ratio) + "_bl" + std::to_string(bl);
     });
 
 // Test supported matmul kernels with clamping support.
@@ -368,15 +403,86 @@ INSTANTIATE_TEST_SUITE_P(
             MatrixPortion(0, 0.75, 1, 1),  // Rightmost portion.
             MatrixPortion(0, 0.5, 1, 0.8)  // Somewhere Middle
             ),
-        testing::ValuesIn(std::initializer_list<float>({1.0f, 0.9f, 0.5f}))),  // clamp_keep_ratio
-    [](const auto& info) {
-        const auto variant_idx = std::get<0>(info.param);
-        const std::string name{variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p.at(variant_idx).variant.ukernel.name};
-        const auto shape = std::get<MatMulShape>(info.param);
-        const auto portion = std::get<2>(info.param);
-        const auto clamp_keep_ratio = std::get<3>(info.param);
+        testing::ValuesIn(std::initializer_list<float>({1.0f, 0.9f, 0.5f})),  // clamp_keep_ratio
+        testing::Values(32), testing::Values(false)),
+    testing::PrintToStringParamName());
 
-        return test_description(name, shape, portion, true, clamp_keep_ratio);
-    });
+// Test kernels with variable block length support
+static constexpr std::array shapes_k32{
+    MatMulShape{1, 2, 32},     //
+    MatMulShape{1, 40, 32},    //
+    MatMulShape{1, 33, 32},    //
+    MatMulShape{15, 32, 32},   //
+    MatMulShape{32, 64, 64},   //
+    MatMulShape{16, 32, 64},   //
+    MatMulShape{32, 64, 128},  //
+    MatMulShape{32, 64, 256},  //
+    MatMulShape{77, 99, 256},
+};
+static constexpr std::array shapes_k64{
+    MatMulShape{32, 64, 64},   //
+    MatMulShape{16, 32, 64},   //
+    MatMulShape{8, 32, 64},    //
+    MatMulShape{77, 99, 64},   //
+    MatMulShape{32, 64, 128},  //
+    MatMulShape{16, 32, 128},  //
+    MatMulShape{32, 64, 256},  //
+    MatMulShape{77, 99, 256},
+};
+static constexpr std::array shapes_k128{
+    MatMulShape{32, 64, 128},  //
+    MatMulShape{16, 32, 128},  //
+    MatMulShape{8, 32, 128},   //
+    MatMulShape{77, 99, 128},  //
+    MatMulShape{32, 64, 256},  //
+    MatMulShape{77, 99, 256},
+};
+static constexpr std::array shapes_k256{
+    MatMulShape{32, 64, 256},  //
+    MatMulShape{16, 32, 256},  //
+    MatMulShape{8, 32, 256},   //
+    MatMulShape{77, 99, 256},
+};
+static constexpr std::array portions{
+    MatrixPortion(0, 0, 1, 1),     // Full matrix.
+    MatrixPortion(0, 0, 1, 0.25),  // Leftmost portion.
+    MatrixPortion(0, 0.75, 1, 1),  // Rightmost portion.
+    MatrixPortion(0, 0.5, 1, 0.8)  // Somewhere Middle
+};
+INSTANTIATE_TEST_SUITE_P(
+    MatMulVariableBL_bl32, MatMulTest_f32_qsi8d32p_qsi4c32p,
+    testing::Combine(
+        testing::Range<size_t>(0, variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p_variable_bl.size()),
+        testing::ValuesIn(shapes_k32), testing::ValuesIn(portions),
+        testing::ValuesIn(std::initializer_list<float>{1.0F}),  // We keep 100% of values - no clamping
+        testing::Values(32), testing::Values(true)),
+    testing::PrintToStringParamName());
+
+INSTANTIATE_TEST_SUITE_P(
+    MatMulVariableBL_bl64, MatMulTest_f32_qsi8d32p_qsi4c32p,
+    testing::Combine(
+        testing::Range<size_t>(0, variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p_variable_bl.size()),
+        testing::ValuesIn(shapes_k64), testing::ValuesIn(portions),
+        testing::ValuesIn(std::initializer_list<float>{1.0F}),  // We keep 100% of values - no clamping
+        testing::Values(64), testing::Values(true)),
+    testing::PrintToStringParamName());
+
+INSTANTIATE_TEST_SUITE_P(
+    MatMulVariableBL_bl128, MatMulTest_f32_qsi8d32p_qsi4c32p,
+    testing::Combine(
+        testing::Range<size_t>(0, variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p_variable_bl.size()),
+        testing::ValuesIn(shapes_k128), testing::ValuesIn(portions),
+        testing::ValuesIn(std::initializer_list<float>{1.0F}),  // We keep 100% of values - no clamping
+        testing::Values(128), testing::Values(true)),
+    testing::PrintToStringParamName());
+
+INSTANTIATE_TEST_SUITE_P(
+    MatMulVariableBL_bl256, MatMulTest_f32_qsi8d32p_qsi4c32p,
+    testing::Combine(
+        testing::Range<size_t>(0, variants_kai_matmul_clamp_f32_qsi8d32p_qsi4c32p_variable_bl.size()),
+        testing::ValuesIn(shapes_k256), testing::ValuesIn(portions),
+        testing::ValuesIn(std::initializer_list<float>{1.0F}),  // We keep 100% of values - no clamping
+        testing::Values(256), testing::Values(true)),
+    testing::PrintToStringParamName());
 
 }  // namespace kai::test